Organize internal common functions
