--- src/main/kotlin/com/fasterxml/jackson/module/kotlin/Types.kt
@@ -1,6 +1,0 @@
-package com.fasterxml.jackson.module.kotlin
-
-import kotlin.reflect.KType
-import kotlin.reflect.jvm.jvmErasure
-
-internal fun KType.erasedType(): Class<out Any> = this.jvmErasure.java
